Fix warning about Airflow version in venv not matching the Astro Runtime version
If you need access to connections inside the virtual environment the only way
right now to currently do that is install Airflow in it. However when launched
in Astro Runtime it failed wit this message:

> The version of Airflow installed for the /home/astro/.pyenv/versions/env1/bin/python(2.7.2)
> is different than the runtime Airflow version: 2.7.2+astro.3

This fixes that by making the check ignore the post release part.

We also needed to force plugins and providers to be lazy, other wise it
conflicts with the non-lazy plugins we have in Astro