Correctly handle producing multiple Venvs for the same python version
I had about 80% of the approach implemented from day one - we already
checked a map to see if we'd already seen that version. The problem was
we never set anything in it so it was always empty.

This also tidies up the tests a bit, and removes the needless link to
`/usr/local/include/python3.7m` for >=Py3.8 (which doesn't need it)

I used `coreos/go-semver` for Semver parsing as that was already in our
dep tree